followed the modification of the original version.
preparing for pull request.

preparing for pull request.
